chore: update imports and move types to src root
