+++ packages/oasis-actions/src/operations/aave/borrow.ts
@@ -3,7 +3,7 @@ import BigNumber from 'bignumber.js'
 import * as actions from '../../actions'
 import { ADDRESSES } from '../../helpers/addresses'
 import { OPERATION_NAMES } from '../../helpers/constants'
-import { Address } from '../../strategies/types/StrategyParams'
+import { Address } from '../../strategies/types'
 
 export interface BorrowArgs {
   borrowToken: Address
