+++ packages/oasis-actions/src/strategies/aave/depositBorrow.ts
@@ -9,9 +9,8 @@ import * as operations from '../../operations'
 import { AAVEStrategyAddresses } from '../../operations/aave/addresses'
 import { BorrowArgs } from '../../operations/aave/borrow'
 import { DepositArgs } from '../../operations/aave/deposit'
-import { IPositionTransition, SwapData } from '../types'
+import { Address, IPositionTransition, IPositionTransitionDependencies, SwapData } from '../types'
 import { AAVETokens } from '../types/aave'
-import { Address, IPositionTransitionDependencies } from '../types/StrategyParams'
 
 interface DepositBorrowArgs {
   entryToken?: { amountInBaseUnit: BigNumber; symbol: AAVETokens; precision?: number }
