+++ packages/oasis-actions/src/strategies/aave/open.ts
@@ -11,10 +11,10 @@ import { RiskRatio } from '../../helpers/calculations/RiskRatio'
 import { TYPICAL_PRECISION, ZERO } from '../../helpers/constants'
 import * as operations from '../../operations'
 import { AAVEStrategyAddresses } from '../../operations/aave/addresses'
+import { Address } from '../types'
 import { AAVETokens } from '../types/aave/tokens'
 import { IPositionTransition } from '../types/IPositionTransition'
 import { PositionType } from '../types/PositionType'
-import { Address } from '../types/StrategyParams'
 import { SwapData } from '../types/SwapData'
 import { getCurrentPosition } from './getCurrentPosition'
