+++ test/aave/Open.test.ts
@@ -13,8 +13,8 @@ import {
 } from '@oasisdex/oasis-actions/src'
 import aavePriceOracleABI from '@oasisdex/oasis-actions/src/abi/aavePriceOracle.json'
 import { amountFromWei } from '@oasisdex/oasis-actions/src/helpers'
+import { Address } from '@oasisdex/oasis-actions/src/strategies/types'
 import { PositionType } from '@oasisdex/oasis-actions/src/strategies/types/PositionType'
-import { Address } from '@oasisdex/oasis-actions/src/strategies/types/StrategyParams'
 import BigNumber from 'bignumber.js'
 import { expect } from 'chai'
 import { Contract, ContractReceipt, ethers, Signer } from 'ethers'
